nits, cleanup
