+++ crates/stages/Cargo.toml
@@ -13,6 +13,8 @@ reth-primitives = { path = "../primitives" }
 reth-interfaces = { path = "../interfaces" }
 reth-executor = { path = "../executor" }
 reth-rlp = { path = "../common/rlp" }
+reth-db = { path = "../storage/db" }
+reth-provider = { path = "../storage/provider"}
 
 #async
 tokio = { version = "1.21.2", features = ["sync"] }
@@ -28,9 +30,11 @@ itertools = "0.10.5"
 rayon = "1.6.0"
 
 [dev-dependencies]
-reth-db = { path = "../storage/db", features = ["test-utils","mdbx"] }
+# reth
+reth-db = { path = "../storage/db", features = ["test-utils", "mdbx"] }
 reth-interfaces = { path = "../interfaces", features = ["test-utils"] }
 reth-bodies-downloaders = { path = "../net/bodies-downloaders" }
+
 # TODO(onbjerg): We only need this for [BlockBody]
 reth-eth-wire = { path = "../net/eth-wire" }
 reth-headers-downloaders = { path = "../net/headers-downloaders" }
