+++ crates/stages/src/stage.rs
@@ -1,6 +1,6 @@
 use crate::{db::StageDB, error::StageError, id::StageId};
 use async_trait::async_trait;
-use reth_interfaces::db::Database;
+use reth_db::database::Database;
 use reth_primitives::BlockNumber;
 
 /// Stage execution input, see [Stage::execute].
