+++ crates/storage/db/src/tables/utils.rs
@@ -1,3 +1,4 @@
+//! Small database table utilities and helper functions
 use crate::{
     table::{Decode, Decompress, Table},
     Error,
