+++ crates/storage/provider/src/lib.rs
@@ -19,6 +19,9 @@ pub use block::{
     get_cumulative_tx_count_by_hash, insert_canonical_block, BlockProvider, ChainInfo,
     HeaderProvider,
 };
-pub use db_provider::{self as db, ProviderImpl};
+pub use db_provider::{
+    self as db, ProviderImpl, StateProviderImplHistory, StateProviderImplLatest,
+    StateProviderImplRefHistory, StateProviderImplRefLatest,
+};
 pub use reth_interfaces::provider::Error;
 pub use state::{AccountProvider, StateProvider, StateProviderFactory};
